refactor(MissionStepTest): 테스트 코드 추가(사단계, 오단계)
